refactor: separate bookRating in a diferents componentes
